Beta: add breadcrumb link back to main selection screen (#20394)
When navigating to a plugin to manage, users may be confused on how to
get back home to the plugin selection screen. They could click the
Jetpack Beta Tester logo in the header, but that might not be fullly
apprent. This PR adds a small breadcrumb link to get back home.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1045900519