Beta plugin: Detect when master needs an update (#20415)
We detected when a PR has been merged and needs updating to master, and
when a PR has an update, and when an RC has an update. But we forgot to
check for when master has been updated.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1053253061