mirrors: Add auto-release action (#21335)
* mirrors: Add auto-release action

This will allow for automatic creation of a GitHub release, based on the
changelog, when a release version is tagged (probably by the
auto-tagger).

The created release will have an artifact attached with a standardized
zip file, suitable for e.g. uploading as a plugin into WordPress.

* update composer.lock

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1318143942