Premium Content Block: Added a "Remove block and keep content" and fixed to group transformation (#22691)
* Fixed Premium Content transform to group and added a "Remove block and keep content" menu item for Premium Content Block.

The Settings menu item acts like an alias to group transform by using the same function that is used to perform the transformation. This can be helpful when trying to remove the block but also keeping the content.

* Added changelogger

* Fixed duplicate nested core/group issue.

* Changed the function declaration and added comment.

* Make "Remove block and keep content" string translatable.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1828488897