Form block: Add color & spacing design tool (#22465)
* [not verified] Add Color, Spacing and Typography features

* [not verified] Add block deprecation - v2 (`save()` function)

This commit specifies the previous version of the `save()` function of the block to maintain backward compatibility.

* [not verified] Add block deprecation - v2 (old `supports` object)

This commit specifies the previous version of the `supports` object of the block - to maintain backward compatibility.

* [not verified] Fix attributes of the deprecated version v2

This change makes sure that the block attributes are carried over when migrating from deprecated block version v2.

* [not verified] Update fixtures

The commit:
- adds new fixture files that represent the deprecated version v2 of the block
- regenerates current fixture files
- updates `.json` and `.serialized.html` files of the deprecated version v1 of the block - to match it with the updated block markdown

* [not verified] Regenerate deprecated fixtures.

* [not verified] Remove `Typography` from supported design tools

The Form block has many different types of inputs, hence it is not realistic to expect a single typography option to manage all those inputs properly.

* Add margin support.

* Add align design tool.

* Update changelog.

* Removed align tool.

* Update changelog.

* Update changelogs

Co-authored-by: Yan Sern <yan.sern@automattic.com>
Co-authored-by: Yan Sern <yansern@yan.sr>
Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1831637250